Move EnumerableT to Transformers library
