+++ Source/Tests/FunicularSwitch.Generators.Consumer/TransformerSpecs.cs
@@ -166,24 +166,6 @@ public static class Writer
     }
 }
 
-public static class EnumerableM
-{
-    public static IEnumerable<B> Bind<A, B>(this IEnumerable<A> ma, Func<A, IEnumerable<B>> fn) =>
-        ma.SelectMany(fn);
-
-    public static IEnumerable<A> Yield<A>(A a) => [a];
-}
-
-[MonadTransformer(typeof(EnumerableM))]
-public static class EnumerableT
-{
-    public static Monad<IEnumerable<B>> BindT<A, B>(Monad<IEnumerable<A>> ma, Func<A, Monad<IEnumerable<B>>> fn) =>
-        ma.Bind(xs => xs.Aggregate(
-            ma.Return<IEnumerable<B>>([]),
-            (cur, acc) => cur.Bind(xs_ =>
-                fn(acc).Map(y => (IEnumerable<B>) [..xs_, ..y]))));
-}
-
 [TransformMonad(typeof(Writer), typeof(ResultT))]
 public readonly partial record struct WriterResult<A>;
